+++ src/Button/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { Button, ButtonProps } from './Button';
+
+export { Button };
+export type { ButtonProps };
+export default withTheme(Button, 'Button');
