+++ src/ButtonGroup/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { ButtonGroup, ButtonGroupProps } from './ButtonGroup';
+
+export { ButtonGroup };
+export type { ButtonGroupProps };
+export default withTheme(ButtonGroup, 'ButtonGroup');
