+++ src/Checkbox/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { CheckBox, CheckBoxProps } from './CheckBox';
+
+export { CheckBox };
+export type { CheckBoxProps };
+export default withTheme(CheckBox, 'Checkbox');
