+++ src/FAB/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { FAB, FABProps } from './FAB';
+
+export { FAB };
+export type { FABProps };
+export default withTheme(FAB, 'FAB');
