+++ src/Slider/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { Slider, SliderProps } from './Slider';
+
+export { Slider };
+export type { SliderProps };
+export default withTheme(Slider, 'Slider');
