+++ src/SpeedDial/index.tsx
@@ -1,0 +1,7 @@
+import { withTheme } from '../config';
+import { SpeedDial, SpeedDialProps } from './SpeedDial';
+import { SpeedDialActionProps } from './components/SpeedDialAction';
+
+export { SpeedDial };
+export type { SpeedDialProps, SpeedDialActionProps };
+export default withTheme(SpeedDial, 'SpeedDial');
