+++ src/config/theme.tsx
@@ -4,10 +4,10 @@ import { AccessoryProps } from '../avatar/Accessory';
 import { AvatarProps } from '../avatar/Avatar';
 import { BadgeProps } from '../badge/Badge';
 import { BottomSheetProps } from '../BottomSheet';
-import { ButtonProps } from '../buttons/Button';
-import { ButtonGroupProps } from '../buttons/ButtonGroup';
+import { ButtonProps } from '../Button';
+import { ButtonGroupProps } from '../ButtonGroup';
 import { CardProps } from '../card/Card';
-import { CheckBoxProps } from '../checkbox/CheckBox';
+import { CheckBoxProps } from '../Checkbox';
 import { DividerProps } from '../Divider';
 import { HeaderProps } from '../Header';
 import { IconProps } from '../icons/Icon';
@@ -17,7 +17,7 @@ import { ListItemProps } from '../list/ListItemBase';
 import { OverlayProps } from '../overlay/Overlay';
 import { PricingCardProps } from '../pricing/PricingCard';
 import { SearchBarProps } from '../searchbar/SearchBar';
-import { SliderProps } from '../slider/Slider';
+import { SliderProps } from '../Slider';
 import { SocialIconProps } from '../SocialIcon';
 import { TextProps } from '../Text';
 import { TileProps } from '../tile/Tile';
@@ -26,9 +26,10 @@ import { SwitchProps } from '../Switch';
 import { ListItemAccordionProps } from '../list/ListItemAccordion';
 import { TabItemProps, TabProps } from '../tab/Tab';
 import { TabViewProps } from '../tab/TabView';
-import { FABProps } from '../buttons/FAB';
-import { SpeedDialProps, SpeedDialActionProps } from '../buttons/SpeedDial';
+import { FABProps } from '../FAB';
+import { SpeedDialProps, SpeedDialActionProps } from '../SpeedDial';
 import { LinearProgressProps } from '../linearProgress';
+import { ChipProps } from '../Chip';
 
 import colors, { Colors } from './colors';
 
@@ -45,6 +46,7 @@ export interface FullTheme {
   BottomSheet: Partial<BottomSheetProps>;
   Button: Partial<ButtonProps>;
   ButtonGroup: Partial<ButtonGroupProps>;
+  Chip: Partial<ChipProps>;
   Card: Partial<CardProps>;
   CardDivider: Partial<DividerProps>;
   CardFeaturedSubtitle: Partial<TextProps>;
