+++ src/list/ListItemCheckBox.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
 import { withTheme } from '../config';
-import CheckBox, { CheckBoxProps } from '../checkbox/CheckBox';
+import CheckBox, { CheckBoxProps } from '../Checkbox';
 import { RneFunctionComponent } from '../helpers';
 
 const ListItemCheckBox: RneFunctionComponent<CheckBoxProps> = ({
